feat(nova): open game detail in its own window instead of a bottom sheet
Opening a game raised a BottomSheetDialogFragment of just over 3,000 lines that
mixed sheet plumbing with the Compose content it hosted. This gives the detail its
own Activity and splits the file along the seam that was already there.

The Compose half moves verbatim into NovaGameDetailContent.kt so the ordering the
source guards assert is unchanged. The fragment half becomes NovaGameDetailActivity,
which keeps the constructor lambdas the sheet was built with as properties of the
same name and shape, so the 500-line body that used them needed context and
lifecycle substitutions rather than a rewrite.

The window returns the launch it chose rather than performing it, so the stream
starts from the library after the window is gone. Two payloads travel back, mirroring
the two paths the sheet already had: artwork edits merge into the library list, and
the launched game goes to the launch. Passing only an id and looking the game back up
would have dropped the MangoHUD choice made in that window.

PolarisGame is @Serializable, so the game travels as JSON rather than through a
hand-written mirror of PolarisGameJsonAdapter that could drift from it. The
serialization runtime is already on the APK's runtime classpath via the shared model;
this only puts it on the app's compile classpath.

expandBottomSheet sized whatever the fragment's root view was, which was right for
the detail sheet and incidental for the desktop-Steam dialog that also called it.
That dialog is the only sheet the window still hosts, so the helper now sizes the
content view it is handed.

Behaviour is otherwise unchanged; the cinematic restructure is separate.
